refactor: update validation test command and paths
