+++ .github/workflows/validation.yml
@@ -30,5 +30,5 @@ jobs:
         yarn install --frozen-lockfile
         VALIDATION=$(echo $PR_TITLE | sed -e 's/.*\[\(.*\)\].*/\1/')
         if [ -n "$VALIDATION" ]; then
-          yarn test:validation --validation=$VALIDATION
+          yarn test:validation $VALIDATION
         fi
