+++ STRATEGIES.md
@@ -48,7 +48,7 @@ Add your validation to `src/strategies/validations/index.ts`:
 ### Test your validation
 
 ```bash
-yarn test:validation --validation=my-new-validation
+yarn test:validation my-new-validation
 ```
 
 ## Checklist for new strategies
