+++ test/strategies/unit/validation.test.ts
@@ -1,313 +1,129 @@
-import Validation from '../../../src/strategies/validations/validation';
+import snapshotjs from '@snapshot-labs/snapshot.js';
+import fetch from 'cross-fetch';
+import snapshot from '../../../src/strategies';
 
-class TestValidation extends Validation {}
+const args = process.argv
+  .slice(2)
+  .filter(
+    arg =>
+      !arg.startsWith('-') &&
+      !arg.includes('=') &&
+      !arg.includes('.ts') &&
+      !arg.includes('.js') &&
+      !arg.includes('/') &&
+      arg !== 'test' &&
+      arg !== 'jest'
+  );
+
+const validationArg = args[0] || 'basic';
+
+const id = Object.keys(snapshot.validations).find(
+  validation => validationArg === validation
+);
+if (!id) throw 'Validation not found';
+
+const examples = require(
+  `../../../src/strategies/validations/${id}/examples.json`
+).map((example, index) => ({ index, example }));
 
 describe('Validation', () => {
-  describe('constructor', () => {
-    it('should initialize all properties correctly', () => {
-      const author = '0x1234567890abcdef1234567890abcdef12345678';
-      const space = 'test-space';
-      const network = '1';
-      const snapshot = 123456;
-      const params = { test: 'value' };
+  it(`validate: ${id}`, async () => {
+    await Promise.all(
+      examples.map(async ({ example }) => {
+        const validation = new snapshot.validations[id].validation(
+          example.author,
+          example.space,
+          example.network,
+          example.snapshot,
+          example.params
+        );
 
-      const validation = new TestValidation(
-        author,
-        space,
-        network,
-        snapshot,
-        params
-      );
+        // edge-case when using setTimeout to wait for the Passport API to process a job request
+        if (id === 'passport-gated') jest.useRealTimers();
 
-      expect(validation.author).toBe(author);
-      expect(validation.space).toBe(space);
-      expect(validation.network).toBe(network);
-      expect(validation.snapshot).toBe(snapshot);
-      expect(validation.params).toBe(params);
-      expect(validation.id).toBe('');
-      expect(validation.github).toBe('');
-      expect(validation.version).toBe('');
-      expect(validation.title).toBe('');
-      expect(validation.description).toBe('');
-      expect(validation.supportedProtocols).toEqual(['evm']);
-      expect(validation.hasInnerStrategies).toBe(false);
+        const exampleValidation = await validation.validate();
+        const expectedValue =
+          example.valid !== undefined ? example.valid : true;
+        console.log(
+          `Example ${example.author} -> Expected: ${example.valid} Actual: ${exampleValidation}`
+        );
+
+        expect(exampleValidation).toBe(expectedValue);
+      })
+    );
+  }, 30e3);
+
+  // Check schema is valid with examples.json
+  let schema;
+  try {
+    schema = require(`../../../src/strategies/validations/${id}/schema.json`);
+  } catch (error) {
+    schema = null;
+  }
+  (schema ? it : it.skip)(
+    'Check schema (if available) is valid with examples.json',
+    async () => {
+      expect(typeof schema).toBe('object');
+      for (const { example } of examples) {
+        expect(snapshotjs.utils.validateSchema(schema, example.params)).toBe(
+          true
+        );
+      }
+    }
+  );
+});
+
+describe('All validations', () => {
+  it('All validations should have unique id and title', () => {
+    const ids = Object.keys(snapshot.validations).map(
+      validation => snapshot.validations[validation].id
+    );
+    const titles = Object.keys(snapshot.validations).map(
+      validation => snapshot.validations[validation].title
+    );
+    expect(new Set(ids).size).toBe(ids.length);
+    expect(new Set(titles).size).toBe(titles.length);
+  });
+
+  it('All validations should have examples.json', () => {
+    Object.keys(snapshot.validations).forEach(validation => {
+      expect(
+        require(
+          `../../../src/strategies/validations/${validation}/examples.json`
+        )
+      ).toBeTruthy();
     });
   });
 
-  describe('validate() method', () => {
-    it('should call doValidate with default author when no custom author provided', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      const doValidateSpy = jest.spyOn(validation, 'doValidate' as any);
+  it('All validations should have a non-empty id, title, description, github and version', () => {
+    // Loop through all validations, and check that each has a non-empty id, title, description, github and version
+    Object.keys(snapshot.validations).forEach(validation => {
+      const { id, title, description, github, version } =
+        snapshot.validations[validation];
+      expect(id).toBeTruthy();
+      expect(title).toBeTruthy();
+      expect(description).toBeTruthy();
+      expect(github).toBeTruthy();
+      expect(version).toBeTruthy();
 
-      await validation.validate();
-
-      expect(doValidateSpy).toHaveBeenCalledWith(
-        '0x1234567890abcdef1234567890abcdef12345678'
-      );
-    });
-
-    it('should call doValidate with custom author when provided', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      const customAuthor = '0xabcdef1234567890abcdef1234567890abcdef12';
-      const doValidateSpy = jest.spyOn(validation, 'doValidate' as any);
-
-      await validation.validate(customAuthor);
-
-      expect(doValidateSpy).toHaveBeenCalledWith(customAuthor);
-    });
-
-    it('should return false for invalid custom author address type', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['evm'];
-      const starknetAddress =
-        '0x07f71118e351c02f6EC7099C8CDf93AED66CEd8406E94631cC91637f7D7F203A';
-
-      const result = await validation.validate(starknetAddress);
-      expect(result).toBe(false);
-    });
-
-    it('should return the result from doValidate', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      const mockResult = false;
-      jest.spyOn(validation, 'doValidate' as any).mockResolvedValue(mockResult);
-
-      const result = await validation.validate();
-
-      expect(result).toBe(mockResult);
+      // Check that the version is in the correct format
+      expect(version).toMatch(/^\d+\.\d+\.\d+$/);
     });
   });
+});
 
-  describe('doValidate() method', () => {
-    it('should return true by default in base class', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-
-      const result = await validation.validate();
-      expect(result).toBe(true);
-    });
+describe(`\nOthers:`, () => {
+  it('github in validation should be a valid github username', async () => {
+    const github = snapshot.validations[id].github;
+    expect(typeof github).toBe('string');
+    const githubUserData = await fetch(
+      `https://api.github.com/users/${github}`
+    );
+    const githubUser = await githubUserData.json();
+    expect(githubUser.message).not.toEqual('Not Found');
   });
-
-  describe('validateStrategiesLength()', () => {
-    it('should not throw error when hasInnerStrategies is false', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        { strategies: new Array(10).fill({}) }
-      );
-      validation.hasInnerStrategies = false;
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should not throw error when hasInnerStrategies is true and strategies length is within limit', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        { strategies: new Array(5).fill({}) }
-      );
-      validation.hasInnerStrategies = true;
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should not throw error when hasInnerStrategies is true and strategies length equals limit', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        { strategies: new Array(8).fill({}) }
-      );
-      validation.hasInnerStrategies = true;
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should not throw error when hasInnerStrategies is true and strategies is undefined', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.hasInnerStrategies = true;
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should not throw error when hasInnerStrategies is true and strategies is empty', async () => {
-      const validation = new TestValidation(
-        '0x1234567890abcdef1234567890abcdef12345678',
-        'test-space',
-        '1',
-        123456,
-        { strategies: [] }
-      );
-      validation.hasInnerStrategies = true;
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-  });
-
-  describe('validateAddressType()', () => {
-    const VALID_EVM_ADDRESS = '0x1234567890abcdef1234567890abcdef12345678';
-    const VALID_STARKNET_ADDRESS =
-      '0x07f71118e351c02f6EC7099C8CDf93AED66CEd8406E94631cC91637f7D7F203A';
-    const INVALID_ADDRESS = 'invalidAddress';
-
-    it('should not throw error for valid EVM address when evm protocol is supported', async () => {
-      const validation = new TestValidation(
-        VALID_EVM_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['evm'];
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should not throw error for valid Starknet address when starknet protocol is supported', async () => {
-      const validation = new TestValidation(
-        VALID_STARKNET_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['starknet'];
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should not throw error for valid EVM address when both protocols are supported', async () => {
-      const validation = new TestValidation(
-        VALID_EVM_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['evm', 'starknet'];
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should not throw error for valid Starknet address when both protocols are supported', async () => {
-      const validation = new TestValidation(
-        VALID_STARKNET_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['evm', 'starknet'];
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should not throw error for valid EVM address when only starknet protocol is supported', async () => {
-      const validation = new TestValidation(
-        VALID_EVM_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['starknet'];
-
-      await expect(validation.validate()).resolves.not.toThrow();
-    });
-
-    it('should return false for valid Starknet address when only evm protocol is supported', async () => {
-      const validation = new TestValidation(
-        VALID_STARKNET_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['evm'];
-
-      const result = await validation.validate();
-      expect(result).toBe(false);
-    });
-
-    it('should return false for invalid address when evm protocol is supported', async () => {
-      const validation = new TestValidation(
-        INVALID_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['evm'];
-
-      const result = await validation.validate();
-      expect(result).toBe(false);
-    });
-
-    it('should return false for invalid address when starknet protocol is supported', async () => {
-      const validation = new TestValidation(
-        INVALID_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['starknet'];
-
-      const result = await validation.validate();
-      expect(result).toBe(false);
-    });
-
-    it('should return false for invalid address when both protocols are supported', async () => {
-      const validation = new TestValidation(
-        INVALID_ADDRESS,
-        'test-space',
-        '1',
-        123456,
-        {}
-      );
-      validation.supportedProtocols = ['evm', 'starknet'];
-
-      const result = await validation.validate();
-      expect(result).toBe(false);
-    });
+  it('Version in validation should be a valid string', async () => {
+    const version = snapshot.validations[id].version;
+    expect(typeof version).toBe('string');
   });
 });
